refactor: move some complexity into utils
