+++ src/cli/clear_kinetics_opts.rs
@@ -1,4 +1,4 @@
-use super::super::utils::InputBam;
+use crate::utils::input_bam::InputBam;
 use clap::Args;
 use std::fmt::Debug;
